Created a Sample.Server.Core.csproj to contain all helper classes that could be subsequently moved to the main core infrastructure.
Created the concept of Initializers of DomainEvent Handlers, this is quite ugly because actually it is initializated in Program.cs of sample.server, but the code should be moved to the handler catalog.

Used this initiqalization concepts in NHDenormalizer to reinitialize the QueryModel NHibernate based on version number of the denormalizer.